requirements.txt
streamlit
openai
PyMuPDF
faiss-cpu
sentence-transformers
python-dotenv